+++ src/main/java/de/pauleduardkoenig/dwds_cf/Constants.java
@@ -10,7 +10,7 @@ import java.util.List;
  * @since 1.0.0
  */
 public class Constants {
-    public static final String VERSION = "1.1.1";
+    public static final String VERSION = "1.2.2";
     public static final String PROGRAMM_NAME = "DWDS-SF";
     public static final Image PROGRAMM_ICON = Toolkit.getDefaultToolkit().getImage(ClassLoader.getSystemResource("dwds-sf_logo_256.png"));