+++ src/main/java/de/pauleduardkoenig/dwds_cf/Fetcher.java
@@ -1,15 +1,23 @@
 package de.pauleduardkoenig.dwds_cf;
 
+import com.therazzerapp.milorazlib.FunctionUtils;
+import com.therazzerapp.milorazlib.container.Pair;
 import com.therazzerapp.milorazlib.excel.csv.CSVFile;
+import com.therazzerapp.milorazlib.excel.csv.CSVLoader;
+import com.therazzerapp.milorazlib.http.HTTPSettings;
 import com.therazzerapp.milorazlib.http.HTTPUtils;
 import com.therazzerapp.milorazlib.json.JSONConfigSection;
+import com.therazzerapp.milorazlib.json.JSONLoader;
 import com.therazzerapp.milorazlib.logger.Logging;
+import org.apache.commons.lang3.StringUtils;
 import org.dom4j.DocumentHelper;
 import org.dom4j.io.OutputFormat;
 import org.dom4j.io.XMLWriter;
 
 import java.io.StringWriter;
 import java.nio.charset.StandardCharsets;
+import java.util.LinkedList;
+import java.util.List;
 
 /**
  * <description>
@@ -19,21 +27,66 @@ import java.nio.charset.StandardCharsets;
  */
 public class Fetcher {
 
-
-    public static CSVFile fetchCSV(DWDSRequestBuilder requestBuilder) {
-        return HTTPUtils.getHTTPRequestAsCSV(requestBuilder.setView(DWDSRequestBuilder.ViewType.CSV).build(), ",", "\"", StandardCharsets.UTF_8, DWDS_CF.config.getAsInt(ConfigType.CONNECT_TIMEOUT), DWDS_CF.config.getAsInt(ConfigType.READ_TIMEOUT), DWDS_CF.logger);
+    private static HTTPSettings getHTTPSettings() {
+        List<Pair<String, String>> props = new LinkedList<>();
+        props.add(HTTPSettings.DEFAULT_AGENT);
+        props.add(new Pair<>("Cookie", "dwds_session=" + DWDS_CF.config.getAsString(ConfigType.SESSION_TOKEN)));
+        return new HTTPSettings(props, DWDS_CF.config.getAsInt(ConfigType.CONNECT_TIMEOUT), DWDS_CF.config.getAsInt(ConfigType.READ_TIMEOUT), StandardCharsets.UTF_8);
     }
 
-    public static CSVFile fetchTSV(DWDSRequestBuilder requestBuilder) {
-        return HTTPUtils.getHTTPRequestAsCSV(requestBuilder.setView(DWDSRequestBuilder.ViewType.TSV).build(), "\t", "\"", StandardCharsets.UTF_8, DWDS_CF.config.getAsInt(ConfigType.CONNECT_TIMEOUT), DWDS_CF.config.getAsInt(ConfigType.READ_TIMEOUT), DWDS_CF.logger);
+    private static Object fetchContent(DWDSRequestBuilder requestBuilder) {
+        List<String> content = HTTPUtils.getHTTPRequestContentList(requestBuilder.setView(DWDSRequestBuilder.ViewType.CSV).build(), getHTTPSettings(), DWDS_CF.logger);
+        if (!content.isEmpty() && content.get(0).startsWith("<!DOCTYPE html><html lang=\"de\" itemscope itemtype=")) {
+            if (content.contains("Zugriff auf diese Seite erhalten Sie, wenn Sie im DWDS eingeloggt sind")) {
+                Logging.warning(DWDS_CF.logger, "(Fetcher) No access to corpus: \"" + requestBuilder.getCorpus() + "\", maybe check session token.");
+                return null;
+            }
+            Logging.warning(DWDS_CF.logger, "(Fetcher) Can't connect to corpus: \"" + requestBuilder.getCorpus() + "\", probably timeout.");
+            return null;
+        }
+        if (requestBuilder.getView() == DWDSRequestBuilder.ViewType.CSV || requestBuilder.getView() == DWDSRequestBuilder.ViewType.TSV) {
+            return content;
+        }
+        return StringUtils.join(content, "\n");
+    }
+
+    public static CSVFile fetchCSV(DWDSRequestBuilder requestBuilder, char separator) {
+        List<String> content = FunctionUtils.doWithRetries(() -> {
+            try {
+                //noinspection unchecked
+                return (List<String>) fetchContent(requestBuilder);
+            } catch (ClassCastException ignored) {
+                return null;
+            }
+        }, DWDS_CF.config.getAsInt(ConfigType.FETCH_TRIES));
+        if (content == null || content.size() <= 1) {
+            return null;
+        }
+        return CSVLoader.load(requestBuilder.build(), content, separator, '\"', StandardCharsets.UTF_8, DWDS_CF.logger);
     }
 
     public static JSONConfigSection fetchJSON(DWDSRequestBuilder requestBuilder) {
-        return HTTPUtils.getRequestAsJSON(requestBuilder.setView(DWDSRequestBuilder.ViewType.JSON).build(), true, StandardCharsets.UTF_8, DWDS_CF.config.getAsInt(ConfigType.CONNECT_TIMEOUT), DWDS_CF.config.getAsInt(ConfigType.READ_TIMEOUT), DWDS_CF.logger);
+        String content = FunctionUtils.doWithRetries(() -> {
+            try {
+                return (String) fetchContent(requestBuilder);
+            } catch (ClassCastException ignored) {
+                return null;
+            }
+        }, DWDS_CF.config.getAsInt(ConfigType.FETCH_TRIES));
+        return content == null ? null : JSONLoader.load(content, true, DWDS_CF.logger);
     }
 
     public static String fetchTCF(DWDSRequestBuilder requestBuilder) {
-        String rawXML = HTTPUtils.getHTTPRequestContent(requestBuilder.setView(DWDSRequestBuilder.ViewType.TCF).build(), StandardCharsets.UTF_8, DWDS_CF.config.getAsInt(ConfigType.CONNECT_TIMEOUT), DWDS_CF.config.getAsInt(ConfigType.READ_TIMEOUT), DWDS_CF.logger);
+        String content = FunctionUtils.doWithRetries(() -> {
+            try {
+                return (String) fetchContent(requestBuilder);
+            } catch (ClassCastException ignored) {
+                return null;
+            }
+        }, DWDS_CF.config.getAsInt(ConfigType.FETCH_TRIES));
+        if (content == null) {
+            return null;
+        }
         try {
             OutputFormat format = OutputFormat.createPrettyPrint();
             format.setIndentSize(DWDS_CF.config.getAsInt(ConfigType.TCF_INDENT_LEVEL));
@@ -41,7 +94,7 @@ public class Fetcher {
             format.setIndent(DWDS_CF.config.getAsString(ConfigType.TCF_INDENT));
             format.setEncoding("UTF-8");
 
-            org.dom4j.Document document = DocumentHelper.parseText(rawXML);
+            org.dom4j.Document document = DocumentHelper.parseText(content);
             StringWriter sw = new StringWriter();
             XMLWriter writer = new XMLWriter(sw, format);
             writer.write(document);
@@ -49,6 +102,6 @@ public class Fetcher {
         } catch (Exception e) {
             Logging.stackTrace(DWDS_CF.logger, e);
         }
-        return rawXML;
+        return content;
     }
 }