+++ src/main/java/de/pauleduardkoenig/dwds_cf/gui/GUIChangelog.java
@@ -34,6 +34,16 @@ public class GUIChangelog extends JDialog {
                     <h1 id="changelog">Changelog</h1>
                     <p>All notable changes to this project will be documented in this file.</p>
                     <h2 id="-beta-builds-">[Beta Builds]</h2>
+                    <h2 id="-1-2-2-15-02-2024">[1.2.2] - 15.02.2024</h2>
+                    <h3 id="add">Add</h3>
+                    <ul>
+                    <li>Corpora that need login. Use session cookie &quot;dwds_session&quot; value from browser in settings to get access.</li>
+                    <li>Option to set retry amounts for fetching.</li>
+                    </ul>
+                    <h3 id="fix">Fix</h3>
+                    <ul>
+                    <li>Bug if row was too long</li>
+                    </ul>
                     <h2 id="-1-1-1-14-02-2024">[1.1.1] - 14.02.2024</h2>
                     <h3 id="add">Add</h3>
                     <ul>
@@ -50,6 +60,7 @@ public class GUIChangelog extends JDialog {
                     <ul>
                     <li>Initial Release</li>
                     </ul>
+                    
                 """);
         logPane.setCaretPosition(0);
     }