+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 ## [Beta Builds]
 
+## [1.2.2] - 15.02.2024
+
+### Add
+
+- Corpora that need login. Use session cookie "dwds_session" value from browser in settings to get access.
+- Option to set retry amounts for fetching.
+
+### Fix
+
+- Bug if row was too long
+
 ## [1.1.1] - 14.02.2024
 
 ### Add